added tests for gremium_get
